move flare_checks to separate module
